Fix printing segv
git-svn-id: svn+ssh://svn.abisource.com/svnroot/abiword/trunk@5105 bcba8976-2d24-0410-9c9c-aab3bd5fdfd6

+++ src/af/xap/unix/gnome/xap_UnixGnomePrintGraphics.cpp
@@ -407,9 +407,9 @@ void XAP_UnixGnomePrintGraphics::setColor(UT_RGBColor& clr)
 	xxx_UT_DEBUGMSG(("Dom: setColor\n"));
 
 	gnome_print_setrgbcolor(m_gpc,
-				(int)(m_currentColor.m_red / 255),
-				(int)(m_currentColor.m_grn / 255),
-				(int)(m_currentColor.m_blu / 255));
+							(m_currentColor.m_red / 255.0),
+							(m_currentColor.m_grn / 255.0),
+							(m_currentColor.m_blu / 255.0));
 	
 }
